Use size_t for grid dimensions and indices
To make the code more readable and avoid potential issues like invalid
negative values, I've changed the grid dimensions and indices to use
size_t instead of int.

+++ src/gpu_game_of_life.cuh
@@ -10,7 +10,7 @@
 class GPUGameOfLife : public GameOfLife {
  public:
   // Constructor that initializes a grid with given dimensions
-  GPUGameOfLife(int width, int height);
+  GPUGameOfLife(size_t width, size_t height);
 
   // Destructor to release CUDA resources
   ~GPUGameOfLife() override;
